+++ starwars/src/components/StarWars.css
@@ -1,0 +1,66 @@
+.char-list {
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+  }
+  
+  .character-card {
+    display: flex;
+    width: 30%;
+    min-width: 375px;
+    height: 250px;
+    background-color: #282727;
+    border: 3px solid #9e4f60;
+    border-radius: 10px;
+    margin-bottom: 20px;
+    color: #ddd;
+    box-shadow: 0 10px 20px rgb(0, 0, 0);
+    overflow: hidden;
+  }
+  
+  .card-leftSide {
+    display: flex;
+    flex-direction: column;
+    width: 45%;
+    height: 100%;
+    border-right: 3px solid #9e4f60;
+    overflow: hidden;
+  }
+  
+  .img-wrapper {
+    width: 170px;
+    object-fit: cover;
+    margin: 0;
+  }
+  
+  .char-img {
+    width: 100%;
+    height: 180px;
+    border-bottom: 3px solid #9e4f60;
+  }
+  
+  .character-name {
+    display:flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 2rem;
+  }
+  
+  .card-rightSide {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 55%;
+    height: 100%;
+  }
+  
+  .character-profile {
+    padding: 0;
+  }
+  
+  .info-list {
+    list-style: none;
+    margin-bottom: 10px;
+    font-size: 1.7rem;
+  }
+
